+++ tests/test_sphinx.py
@@ -6,7 +6,7 @@
 # Public License v3. See the LICENSE file or http://www.gnu.org/licenses/.
 
 
-from collections import namedtuple
+import logging
 
 import pytest
 
@@ -16,19 +16,23 @@ from sphinx.util.docutils import docutils_namespace
 
 from rinoh.document import DocumentTree
 from rinoh.frontend.sphinx import (template_from_config, set_document_metadata,
-                                   deprecation_warnings)
+                                   variable_removed_warnings)
 from rinoh.language import IT
-from rinoh.paper import A4, LETTER
+from rinoh.paper import A4
 from rinoh.templates import Book, Article
 
 
+LOGGER = logging.getLogger(__name__)
+
+
 def create_sphinx_app(tmpdir, **confoverrides):
-    app = Sphinx(srcdir=tmpdir.strpath,
-                 confdir=None,
-                 outdir=(tmpdir / 'output').strpath,
-                 doctreedir=(tmpdir / 'doctrees').strpath,
-                 buildername='rinoh',
-                 confoverrides=confoverrides)
+    with docutils_namespace():
+        app = Sphinx(srcdir=tmpdir.strpath,
+                     confdir=None,
+                     outdir=(tmpdir / 'output').strpath,
+                     doctreedir=(tmpdir / 'doctrees').strpath,
+                     buildername='rinoh',
+                     confoverrides=confoverrides)
     return app
 
 
@@ -38,14 +42,6 @@ def create_document(title='A Title', author='Ann Other', docname='a_name'):
     document.settings.author = author
     document.settings.docname = docname
     return document
-
-
-def create_logger():
-    MockLogger = namedtuple('MockLogger', ['warning', 'warnings'])
-    warnings = []
-    def capture_warning(msg):
-        warnings.append(msg)
-    return MockLogger(capture_warning, warnings)
 
 
 CONFIG_DIR = 'confdir'
@@ -61,11 +57,9 @@ def get_contents_page_size(template_configuration):
     return page.get_config_value('page_size', doc)
 
 
-def get_template_cfg(tmpdir, warn=print, **confoverrides):
-    with docutils_namespace():
-        app = create_sphinx_app(tmpdir, **confoverrides)
-        template_cfg = template_from_config(app.config, CONFIG_DIR, warn)
-    return template_cfg
+def get_template_cfg(tmpdir, **confoverrides):
+    app = create_sphinx_app(tmpdir, **confoverrides)
+    return template_from_config(app.config, CONFIG_DIR, LOGGER)
 
 
 def test_sphinx_config_default(tmpdir):
@@ -87,13 +81,12 @@ def test_sphinx_config_language(tmpdir):
     assert template_cfg['language'] == IT
 
 
-def test_sphinx_config_language_not_supported(tmpdir):
-    logger = create_logger()
-    template_cfg = get_template_cfg(tmpdir, warn=logger.warning, language='not_supported')
+def test_sphinx_config_language_not_supported(caplog, tmpdir):
+    with caplog.at_level(logging.WARNING):
+        template_cfg = get_template_cfg(tmpdir, language='not_supported')
+    assert "The language 'not_supported' is not supported" in caplog.text
     assert template_cfg.template == Book
     assert 'language' not in template_cfg
-    the_warning, = logger.warnings
-    assert 'The language "not_supported" is not supported by rinohtype' in the_warning
 
 
 def test_sphinx_config_rinoh_template(tmpdir):
@@ -125,7 +118,7 @@ def test_sphinx_config_rinoh_template_from_filename(tmpdir):
 
 def test_sphinx_set_document_metadata(tmpdir):
     app = create_sphinx_app(tmpdir, release='1.0', rinoh_template='book')
-    template_cfg = template_from_config(app.config, CONFIG_DIR, print)
+    template_cfg = template_from_config(app.config, CONFIG_DIR, LOGGER)
     docutils_tree = create_document()
     rinoh_tree = DocumentTree([])
     rinoh_doc = template_cfg.document(rinoh_tree)
@@ -141,7 +134,7 @@ def test_sphinx_set_document_metadata_subtitle(tmpdir):
     expected_subtitle = 'A subtitle'
     app = create_sphinx_app(tmpdir, rinoh_metadata={
                             'subtitle': expected_subtitle})
-    template_cfg = template_from_config(app.config, CONFIG_DIR, print)
+    template_cfg = template_from_config(app.config, CONFIG_DIR, LOGGER)
     docutil_tree = create_document()
     rinoh_tree = DocumentTree([])
     rinoh_doc = template_cfg.document(rinoh_tree)
@@ -151,8 +144,8 @@ def test_sphinx_set_document_metadata_subtitle(tmpdir):
 
 def test_sphinx_set_document_metadata_logo(tmpdir):
     expected_logo = 'logo.png'
-    app = create_sphinx_app(tmpdir, rinoh_logo="logo.png")
-    template_cfg = template_from_config(app.config, CONFIG_DIR, print)
+    app = create_sphinx_app(tmpdir, rinoh_logo='logo.png')
+    template_cfg = template_from_config(app.config, CONFIG_DIR, LOGGER)
     docutil_tree = create_document()
     rinoh_tree = DocumentTree([])
     rinoh_doc = template_cfg.document(rinoh_tree)
@@ -160,24 +153,22 @@ def test_sphinx_set_document_metadata_logo(tmpdir):
     assert expected_logo == rinoh_doc.metadata['logo']
 
 
-def test_sphinx_default_deprecation_warning(tmpdir):
+def test_sphinx_default_deprecation_warning(caplog, tmpdir):
     app = create_sphinx_app(tmpdir)
-    logger = create_logger()
-    deprecation_warnings(app.config, logger)
-    assert logger.warnings == []
+    with caplog.at_level(logging.WARNING):
+        variable_removed_warnings(app.config, LOGGER)
+    assert caplog.text == ''
 
 
-def test_sphinx_rinoh_stylesheet_deprecation_warning(tmpdir):
+def test_sphinx_rinoh_stylesheet_removed(caplog, tmpdir):
     app = create_sphinx_app(tmpdir, rinoh_stylesheet="sphinx")
-    logger = create_logger()
-    deprecation_warnings(app.config, logger)
-    the_warning, = logger.warnings
-    assert "Support for 'rinoh_stylesheet' has been removed" in the_warning
+    with caplog.at_level(logging.WARNING):
+        variable_removed_warnings(app.config, LOGGER)
+    assert "Support for 'rinoh_stylesheet' has been removed" in caplog.text
 
 
-def test_sphinx_rinoh_paper_size_deprecation_warning(tmpdir):
-    app = create_sphinx_app(tmpdir, rinoh_paper_size="A4")
-    logger = create_logger()
-    deprecation_warnings(app.config, logger)
-    the_warning, = logger.warnings
-    assert "Support for 'rinoh_paper_size' has been removed" in the_warning
+def test_sphinx_rinoh_paper_size_removed(caplog, tmpdir):
+    app = create_sphinx_app(tmpdir, rinoh_paper_size='A4')
+    with caplog.at_level(logging.WARNING):
+        variable_removed_warnings(app.config, LOGGER)
+    assert "Support for 'rinoh_paper_size' has been removed" in caplog.text